fix: image upload code 도파
